+++ packages/wepy-cli/src/compile.js
@@ -247,6 +247,29 @@ export default {
                 });
                 return false;
             }
+        } else if (config.output === 'baidu') {
+            wepyrc.build = wepyrc.build || {};
+            wepyrc.build.baidu = wepyrc.build.baidu || {};
+            wepyrc.build.baidu.dist = wepyrc.build.baidu.dist || 'baidu';
+            wepyrc.build.baidu.src = wepyrc.build.baidu.src || 'src';
+            if (wepyrc.build.baidu.resolve)
+                wepyrc.resolve = Object.assign({}, wepyrc.resolve, wepyrc.build.baidu.resolve);
+            wepyrc.output = 'baidu';
+
+            resolve.init(wepyrc.resolve || {});
+            loader.attach(resolve);
+
+            if (!resolve.getPkg('wepy-baidu')) {
+                util.log('正在尝试安装缺失资源 wepy-baidu，请稍等。', '信息');
+                util.exec(`npm install wepy-baidu --save`).then(d => {
+                    util.log(`已完成安装 wepy-baidu，重新启动编译。`, '完成');
+                    this.build(config);
+                }).catch(e => {
+                    util.log(`安装插件失败：wepy-baidu，请尝试运行命令 "npm install wepy-baidu --save" 进行安装。`, '错误');
+                    console.log(e);
+                });
+                return false;
+            }
         }
 
         return true;
@@ -301,6 +324,13 @@ export default {
                     files = [file];
                 } else {
                     files = refs;
+                }
+                // imported in less / sass.
+                let fullpath = path.join(current, src, file);
+                let cssDeps = cache.getCssDep(fullpath);
+                if (cssDeps.length) {
+                  let cssDepfiles = cssDeps.map(v =>  path.relative(path.join(current, src), v));
+                  files = refs.length === 0 ? cssDepfiles : files.concat(cssDepfiles);
                 }
             } else if (file.indexOf('components') !== -1) { // 是wpy 文件，而且是组件
                 let parents = this.findParents(file);