+++ packages/wepy-cli/src/util.js
@@ -139,7 +139,7 @@ const utils = {
             }
         }
         let lib = com, main = null;
-        if (com.indexOf(path.sep) > 0) {
+        if (com[0] !== '@' && com.indexOf(path.sep) > 0) {
             let sepIndex = com.indexOf(path.sep);
             lib = com.substring(0, sepIndex);
             main = com.substring(sepIndex + 1, com.length);
@@ -147,6 +147,9 @@ const utils = {
         let o = resolve.getMainFile(lib);
         if (o) {
             if (main) {
+                if (o.pkg && o.pkg._activeFields.length) {
+                    main = resolve.resolveSelfFields(o.dir, o.pkg, main) || main;
+                }
                 src = path.join(o.dir, main);
             } else {
                 src = path.join(o.dir, o.file);
@@ -187,6 +190,18 @@ const utils = {
     },
     isArray (obj) {
         return Array.isArray(obj);
+    },
+    isTrue (v) {
+        return v === true;
+    },
+    isFalse (v) {
+        return v === false;
+    },
+    isUndef (v) {
+        return v === undefined || v === null;
+    },
+    isDef (v) {
+        return v !== undefined && v !== null;
     },
     isFile (p) {
         p = (typeof(p) === 'object') ? path.join(p.dir, p.base) : p;
@@ -310,6 +325,14 @@ const utils = {
                 return ' a:' + name;
             });
         }
+
+        // replace wx: to s-
+        if(config.output === 'baidu') {
+          content = content.replace(/\s+wx\:(\w+)/ig, (match, name) => {
+            return ' s-' + name;
+          });
+        }
+
         return content.replace(/<([\w-]+)\s*[\s\S]*?(\/|<\/[\w-]+)>/ig, (tag, tagName) => {
             tagName = tagName.toLowerCase();
             return tag.replace(/\s+:([\w-_]*)([\.\w]*)\s*=/ig, (attr, name, type) => { // replace :param.sync => v-bind:param.sync
@@ -455,7 +478,8 @@ const utils = {
         ext = (ext ? (ext[0] === '.' ? ext : ('.' + ext)) : opath.ext);
         // 第三组件
         if (opath.npm) {
-            relative = path.relative(opath.npm.modulePath, opath.npm.dir);
+            // maybe it's node_modules/moduleA/xxx/yyy/zzz/index.js
+            relative = path.relative(opath.npm.modulePath, opath.dir);
             relative = path.join('npm', relative);
         } else {
             relative = path.relative(path.join(this.currentDir, src), opath.dir);
@@ -517,8 +541,23 @@ const utils = {
         return rst;
     },
     getVersion () {
+        let version;
         let filepath = path.resolve(__dirname, '../package.json');
-        let version = JSON.parse(this.readFile(filepath)).version;
+        try {
+            version = JSON.parse(this.readFile(filepath)).version;
+        } catch (e) {
+            version = '';
+        }
+        return version;
+    },
+    getProjectVersion () {
+        let version;
+        let filepath = path.resolve(this.currentDir, 'package.json');
+        try {
+            version = JSON.parse(this.readFile(filepath)).version;
+        } catch (e) {
+            version = '';
+        }
         return version;
     },
     datetime (date = new Date(), format = 'HH:mm:ss') {
@@ -594,9 +633,11 @@ const utils = {
             msg = msg.replace(/\\/g, '\\\\');
             msg = msg.replace(/\u001b/g, '');
             msg = msg.replace(/\[\d+m/g, '');
+            msg = msg.replace(/`/g, '\\`');
         }
         try {
-            fs.appendFileSync(file, `console.${type}(\`CLI报错：${msg}\`);\r\n`);
+            msg = msg.replace(/\n/g, '\\n').replace(/\"/g, '\\"');
+            fs.appendFileSync(file, `console.${type}("CLI报错：${msg}");\r\n`);
         } catch (e) {
             console.log(e);
         }