+++ packages/wepy-redux/src/connect/index.js
@@ -24,7 +24,7 @@ export default function connect (states, actions) {
             const store = getStore();
             let hasChanged = false;
             Object.keys(states).forEach((k) => {
-                const newV = states[k]();
+                const newV = states[k].call(this);
                 if (this[k] !== newV) {
                     // 不相等
                     this[k] = newV;
@@ -36,11 +36,7 @@ export default function connect (states, actions) {
         return class extends Component {
             constructor () {
                 super();
-                this.computed = Object.assign(this.computed || {}, states, mapState({
-                    $state(state) {
-                        return state;
-                    }
-                }));
+                this.computed = Object.assign(this.computed || {}, states);
                 this.methods = Object.assign(this.methods || {}, actions);
             }
             onLoad() {